backend/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
openai==1.12.0
python-wordpress-xmlrpc==2.3
beautifulsoup4==4.12.3
feedparser==6.0.10
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
httpx==0.26.0
schedule==1.2.1
apscheduler==3.10.4